+++ packages/js/src/ai-content-planner/containers/feature-modal.js
@@ -7,7 +7,7 @@ import { STORE_NAME_AI } from "../../ai-generator/constants";
 
 export default compose( [
 	withSelect( ( select ) => {
-		const { selectFeatureModalStatus, selectIsModalOpen } = select( CONTENT_PLANNER_STORE );
+const { selectFeatureModalStatus, selectIsModalOpen, selectSuggestion } = select( CONTENT_PLANNER_STORE );
 		const content = select( "core/editor" ).getEditedPostContent();
 		const { getIsPremium, selectLink } = select( "yoast-seo/editor" );
 		const { isUsageCountLimitReached } = select( STORE_NAME_AI );
@@ -19,6 +19,7 @@ export default compose( [
 			status: selectFeatureModalStatus(),
 			upsellLink: selectLink( "https://yoa.st/content-planner-approve-modal" ),
 			isUpsell: isUsageCountLimitReached(),
+			selectedSuggestion: selectSuggestion(),
 		};
 	} ),
 	withDispatch( ( dispatch ) => {
